Modularize CloudSync workspace persistence
Separate account binding, workspace projection, recovery scenarios, and their tests behind the existing CloudSync API.
